Made souls float values and added calculation functions to Player

+++ SoulHarvest/upgrades.h
@@ -8,7 +8,7 @@ public:
 
     int amountOwned;
     int cost;
-    int soulsPerSec;
+    float soulsPerSec;
 
 private:
